requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
Django==5.2.2
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
Werkzeug==3.1.3
